Distinctive collocates data

The input data for distinctive collocates analysis in Rajeg (2018, Ch.7). The collocates data are words co-occurring within the span of 4 words to the right and left of the happiness near-synonyms. Following Gevaert (2007, p. 197), the collocates that are part of the studied synonyms are excluded because they may neutralise the distinctive features of the near-synonyms when the goal is to find semantic differences between these near-synonyms.

colloc_input_data

Format

An object of class tbl_df (inherits from tbl, data.frame) with 27989 rows and 2 columns.

Details

The data consists of the following two variables:

synonyms

The ten near-synonyms under studied. Note that only results of the nominalised, ke- -an forms are presented in the thesis, but not the root forms.

collocates

The window-span collocates of the synonyms

References

Rajeg, G. P. W. (2018). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS (PhD Thesis). Monash University. Melbourne, Australia.

Gevaert, C. (2007). The history of ANGER: The lexical field of ANGER from Old to Early Modern English (PhD thesis). Katholieke Universiteit Leuven, Leuven.