Retrieve the conventionalised Lexical Units of a metaphor

The function helps filter out LUs of a metaphor that can be considered as conventional linguistic instances (cf. Vergara Wilson, 2014, for the way to determine CIC, which stands for "conventionalised instances of constructions").

get_cic_meta_lu(metaphor = NULL, df_ttr_out = NULL)

Arguments

metaphor

regular expressions for the relevant metaphor.

df_ttr_out

the tibble as the output of ttr.

Value

A tibble data frame

References

Rajeg, G. P. W. (2019). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS (PhD Thesis). Monash University. Melbourne, Australia. https://doi.org/10.26180/5cac231a97fb1.

Vergara Wilson, D. (2014). Categorization and Constructional Change in Spanish Expressions of "becoming". Leiden: Brill.

Examples

ttr_metaphor <- ttr(df = phd_data_metaphor,
                    schema_var = "metaphors",
                    lexunit_var = "lu",
                    float_digits = 2)

get_cic_meta_lu("possessable", ttr_metaphor)
#> # A tibble: 13 × 4
#>    Lexical_units  Gloss                               N Perc_overall
#>    <chr>          <chr>                           <int>        <dbl>
#>  1 *beri(kan)*    to give                           170        22.7 
#>  2 *dapat(kan).v* to get                            120        16.0 
#>  3 *bawa(kan)*    to bring                           75        10.0 
#>  4 *berbagi*      to share                           71         9.48
#>  5 *peroleh*      to acquire/earn                    43         5.74
#>  6 *miliki*       to have/own                        35         4.67
#>  7 *kehilangan*   to lose                            22         2.94
#>  8 *kembalikan*   to give back; to return            20         2.67
#>  9 *datangkan*    to bring sth.; to cause to come    16         2.14
#> 10 *hadirkan*     to present sth.                    15         2   
#> 11 *bagi.pp*      for sb.                            14         1.87
#> 12 *punya*        to have/own                        13         1.74
#> 13 *bagi(kan).v*  to share                           12         1.6