Retrieve the top-10 lexically diverse metaphors

This function calls the script to output the top-10 metaphors that have high index of lexical creativity in their linguistic manifestation. The lexical creativity is derived from the type-per-token ratio (TTR) measure (Rajeg, 2018, Chapter 6). The alternative function with shorter name is get_creative_metaphors.

get_lexically_diverse_metaphors(
  df_ttr_out = NULL,
  min_token = 3,
  top_n_limit = 10
)

Arguments

df_ttr_out

the tibble as the output of ttr.

min_token

minimum token frequency of the metaphors.

top_n_limit

limit for presenting the top-n most lexically diverse metaphors.

Value

A tibble data frame.

References

Rajeg, G. P. W. (2018). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS (PhD Thesis). Monash University. Melbourne, Australia.

Examples

# get the ttr table
ttr_metaphor <- ttr(phd_data_metaphor,
                    schema_var = "metaphors",
                    lexunit_var = "lu",
                    float_digits = 2)

# get the lexically diverse metaphors
get_lexically_diverse_metaphors(df_ttr_out = ttr_metaphor,
                                min_token = 3,
                                top_n_limit = 10)
#> # A tibble: 10 × 6
#>    metaphors              token type_lu perc_token perc_type_lu type_per_token_…
#>    <chr>                  <int>   <int>      <dbl>        <dbl>            <dbl>
#>  1 happiness is a harmfu…    11      11       0.3          1.37            100  
#>  2 happiness is drugs         6       6       0.16         0.75            100  
#>  3 happiness is a moved …     4       4       0.11         0.5             100  
#>  4 happiness is a treatm…     3       3       0.08         0.37            100  
#>  5 happiness is an accom…     3       3       0.08         0.37            100  
#>  6 happiness is being so…     9       8       0.25         1                88.9
#>  7 happiness is a resour…     8       7       0.22         0.87             87.5
#>  8 happiness is impedime…     6       5       0.16         0.62             83.3
#>  9 happiness is a deceiv…    20      16       0.55         1.99             80  
#> 10 happiness is an adver…    24      19       0.66         2.36             79.2