Retrieve the (top-n most frequent) Lexical Units evoking the source frames of a metaphor

The function is designed to easily retrieve a tibble data frame of the top-n most frequent frame-evoking Lexical Units (LU) of the prominent metaphors discussed in Rajeg (2019, Chapter 5 and 6).

get_lu_frame_table(
  metaphor = NULL,
  word = NULL,
  print_all = FALSE,
  top_n = FALSE,
  limit = 10,
  df = NULL,
  metaphor_var = "metaphors",
  frame_var = "source_frames",
  synonym_var = "synonyms",
  lexunit_var = "lu",
  lexunit_gloss_var = "lu_gloss"
)

Arguments

metaphor

regular expressions for the relevant metaphor.

word

character strings of regular expressions for the HAPPINESS synonyms.

print_all

logical; print all output

top_n

logical; limit to the top-n LUs

limit

integer; maximum rows to be printed from the top-n LUs

df

the data frame (phd_data_metaphor).

metaphor_var

character string for the column name of the metaphor variable (i.e., "metaphors").

frame_var

character string for the column name of the source frame variable (i.e., "source_frames").

synonym_var

character string for the column name of the synonyms variable (i.e., "synonyms").

lexunit_var

character string for the column name of the lexical unit variable (i.e., "lu").

lexunit_gloss_var

character string for the column name of the gloss of the lexical units (i.e., "lu_gloss").

Value

A tibble/data frame of the descriptive statistics of the LUs and the frames, including the gloss of the LUs.

References

Rajeg, G. P. W. (2019). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS (PhD Thesis). Monash University. Melbourne, Australia. https://doi.org/10.26180/5cac231a97fb1.

Examples

# for the aggregated data
get_lu_frame_table(metaphor = "possessable",
                   top_n = TRUE,
                   limit = 20,
                   df = phd_data_metaphor)
#> # A tibble: 21 × 6
#>    Lexical_units  Frames          Glosses          N Perc_by_frames Perc_overall
#>    <chr>          <chr>           <chr>        <int>          <dbl>        <dbl>
#>  1 *beri(kan)*    giving          to give        170          75.2         22.7 
#>  2 *dapat(kan).v* gain possession to get         120          65.9         16.0 
#>  3 *bawa(kan)*    caused motion   to bring        75          64.7         10.0 
#>  4 *berbagi*      dispersal       to share        71          69.6          9.48
#>  5 *peroleh*      gain possession to acquire/…    43          23.6          5.74
#>  6 *miliki*       possession      to have/own     35          50            4.67
#>  7 *kehilangan*   lose possession to lose         22          81.5          2.94
#>  8 *kembalikan*   giving          to give bac…    20           8.85         2.67
#>  9 *datangkan*    caused motion   to bring st…    16          13.8          2.14
#> 10 *hadirkan*     caused motion   to present …    15          12.9          2   
#> 11 *bagi.pp*      giving          for sb.         14           6.19         1.87
#> 12 *punya*        possession      to have/own     13          18.6          1.74
#> 13 *bagi(kan).v*  dispersal       to share        12          11.8          1.6 
#> 14 *tawarkan*     offering        to offer         9         100            1.2 
#> 15 *pembawa*      caused motion   bringer; ca…     8           6.9          1.07
#> 16 *terima*       gain possession to receive       8           4.4          1.07
#> 17 *milik*        possession      one's posse…     8          11.4          1.07
#> 18 *punyai*       possession      to have/own      8          11.4          1.07
#> 19 *tebar(kan)*   dispersal       to spread s…     7           6.86         0.93
#> 20 *rampas*       theft           to seize         6          35.3          0.8 
#> 21 *renggut*      theft           to snatch/g…     6          35.3          0.8 

# for a specific synonym
get_lu_frame_table(metaphor = "possessable",
                   word = "^bahagia$",
                   top_n = TRUE,
                   limit = 20,
                   df = phd_data_metaphor)
#> # A tibble: 19 × 6
#>    Lexical_units  Frames          Glosses          N Perc_by_frames Perc_overall
#>    <chr>          <chr>           <chr>        <int>          <dbl>        <dbl>
#>  1 *beri(kan)*    giving          to give         11          78.6            22
#>  2 *bawa(kan)*    caused motion   to bring         6          54.6            12
#>  3 *berbagi*      dispersal       to share         6          66.7            12
#>  4 *dapat(kan).v* gain possession to get           5          50              10
#>  5 *punya*        possession      to have/own      3          60               6
#>  6 *hadirkan*     caused motion   to present …     2          18.2             4
#>  7 *pembawa*      caused motion   bringer; ca…     2          18.2             4
#>  8 *peroleh*      gain possession to acquire/…     2          20               4
#>  9 *terima*       gain possession to receive       2          20               4
#> 10 *karuniai*     giving          to grant         2          14.3             4
#> 11 *datangkan*    caused motion   to bring st…     1           9.09            2
#> 12 *bertebaran*   dispersal       to be scatt…     1          11.1             2
#> 13 *hamparkan*    dispersal       to spread o…     1          11.1             2
#> 14 *tebar(kan)*   dispersal       to spread s…     1          11.1             2
#> 15 *sita.v*       gain possession to seize po…     1          10               2
#> 16 *serahkan*     giving          to hand over     1           7.14            2
#> 17 *bekal*        possession      victuals         1          20               2
#> 18 *miliki*       possession      to have/own      1          20               2
#> 19 *rampas*       theft           to seize         1         100               2