Retrieve the frequency of Lexical units and the evoked submappings

The function generates a table of frequency and percentages of Lexical units evoking certain submappings of a metaphor.

get_lu_submappings_table(
  metaphor = NULL,
  word = NULL,
  print_all = FALSE,
  top_n = FALSE,
  limit = 10,
  df = NULL,
  metaphor_var = "metaphors",
  submet_var = "submappings",
  synonym_var = "synonyms",
  lexunit_var = "lu",
  lexunit_gloss_var = "lu_gloss"
)

Arguments

metaphor
regular expressions for the relevant metaphor.

word
character strings of regular expressions for the HAPPINESS synonyms.

print_all
logical; print all output

top_n
logical; limit to the top-n LUs

limit
integer; maximum rows to be printed from the top-n LUs

df
the data frame (phd_data_metaphor).

metaphor_var
character string for the column name of the metaphor variable (i.e., "metaphors").

submet_var
character string for the column name of the submapping variable (i.e., "submappings").

synonym_var
character string for the column name of the synonyms variable (i.e., "synonyms").

lexunit_var
character string for the column name of the lexical unit variable (i.e., "lu").

lexunit_gloss_var
character string for the column name of the gloss of the lexical units (i.e., "lu_gloss").

Value

A tibble containing the submappings, LUs and the glosses, frequency of each LUs of the metaphor, percentage of each LUs in the metaphor, and percentage of each LUs by the evoked submappings.

References

Rajeg, G. P. W. (2018). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS. PhD Thesis. Monash University. Melbourne, Australia.

Examples

# for all synonyms
get_lu_submappings_table(metaphor = "liquid in a container",
                         df = phd_data_metaphor)

#> # A tibble: 37 x 6
#>    submappings       lu     lu_gloss        n perc_expr_overa… perc_expr_by_sub…
#>    <chr>             <chr>  <chr>       <int>            <dbl>             <dbl>
#>  1 expression of ha… terpa… to be spur…    43            27.6                 43
#>  2 expression of ha… luapan overflow       25            16.0                 25
#>  3 expression of ha… panca… to spurt s…     9             5.77                 9
#>  4 expression of ha… salur… to funnel …     4             2.56                 4
#>  5 expression of ha… panca… a spurting…     3             1.92                 3
#>  6 expression of ha… curah… outpouring      2             1.28                 2
#>  7 expression of ha… tertu… to be pour…     2             1.28                 2
#>  8 expression of ha… tuang… to pour ou…     2             1.28                 2
#>  9 expression of ha… air b… flood           1             0.64                 1
#> 10 expression of ha… alirk… to flow an…     1             0.64                 1
#> # … with 27 more rows

# for \emph{kegembiraan} 'joy'
get_lu_submappings_table(metaphor = "liquid in a container",
                         word = "kegembiraan",
                         df = phd_data_metaphor)

#> # A tibble: 12 x 7
#>    synonyms  submappings  lu    lu_gloss     n perc_expr_overa… perc_expr_by_su…
#>    <chr>     <chr>        <chr> <chr>    <int>            <dbl>            <dbl>
#>  1 kegembir… expression … luap… overflow    21            31.3             55.3 
#>  2 kegembir… expression … terp… to be s…    10            14.9             26.3 
#>  3 kegembir… expression … tuan… to pour…     2             2.99             5.26
#>  4 kegembir… expression … air … flood        1             1.49             2.63
#>  5 kegembir… expression … lamp… to gush…     1             1.49             2.63
#>  6 kegembir… expression … memb… to over…     1             1.49             2.63
#>  7 kegembir… expression … panc… to spur…     1             1.49             2.63
#>  8 kegembir… expression … tert… to be p…     1             1.49             2.63
#>  9 kegembir… high intens… limp… to brim…     1             1.49           100   
#> 10 kegembir… intensified… luap… to boil…    19            28.4             76   
#> 11 kegembir… intensified… melu… to boil…     6             8.96            24   
#> 12 kegembir… preventing … bend… to dam …     3             4.48           100