Retrieve the (top-n most frequent) Lexical Units of a metaphor

The function is designed to easily retrieve a tibble data frame of the top-n most frequent Lexical Units (LU) of the prominent metaphors discussed in Rajeg (2019, Chapter 5 and 6).

get_lu_table(
  metaphor = "regular expressions",
  top_n_only = TRUE,
  top_n_limit = 10,
  df = NULL,
  submapping_perc = FALSE,
  incl_submappings = FALSE
)

Arguments

metaphor
regular expressions for the relevant metaphor.

top_n_only
logical; the default is TRUE with the number specified in top_n_limit argument. If FALSE, it prints all the LU for the metaphor.

top_n_limit
integer; how many most frequent LU should be printed? The default is 10.

df
the data frame for the thesis (phd_data_metaphor). This argument is not put first because the main attention when using the function is on varying the metaphor (hence, coming first) rather than on varying the database.

submapping_perc
logical; wheter to include the percentages of the LU by submappings. If TRUE, the function only print out the percentages of the LU by submappings. To include the submapping types, specify the incl_submappings argument to TRUE as well (the default is FALSE).

incl_submappings
logical; whether to include the submappings inferred from the LU (TRUE) or not (FALSE -- the default).

Value

A tbl_df.

References

Rajeg, G. P. W. (2019). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS. PhD Thesis. Monash University. Melbourne, Australia. https://doi.org/10.26180/5cac231a97fb1.

Examples

get_lu_table(metaphor = "desired goal$", df = phd_data_metaphor)

#> # A tibble: 10 x 4
#>    Lexical_units Gloss                                   N Perc_overall
#>    <chr>         <chr>                               <int>        <dbl>
#>  1 *cari*        to search/look for                     74        25.3 
#>  2 *capai*       to reach                               51        17.4 
#>  3 *raih*        to pull/reach sth. towards the body    27         9.22
#>  4 *kejar*       to chase                               24         8.19
#>  5 *temukan*     to find                                24         8.19
#>  6 *tujuan*      destination/goal                       13         4.44
#>  7 *menuju*      to head to                              9         3.07
#>  8 *jalan*       way                                     8         2.73
#>  9 *kunci*       key                                     8         2.73
#> 10 *gapai*       to reach out                            6         2.05