Attracted items from MDCA

A function to retrieve list of distinctive or attracted items for a construction from the results of mdca.

mdca_attr(
  df = NULL,
  filter_by = "cxn",
  cxn_var = "synonyms",
  cxn_type = NULL,
  coll_var = NULL,
  coll_type = NULL,
  min_assocstr = 2
)

Arguments

df

output data frame from mdca.

filter_by

character string indicating whether the output of mdca is filtered according to the construction variable (i.e. "cxn") or the collexemes/collocates variable (i.e. "colloc"). If it is left NULL (the default), the data is filtered according to the value specified in the min_assocstr argument.

cxn_var

character string for the column name of the construction variable (i.e., "synonyms").

cxn_type

character strings of regular expressions for the HAPPINESS synonyms.

coll_var

character string for the column name of the metaphor variable (i.e., "metaphors").

coll_type

character strings of regular expressions for the conceptual metaphors.

min_assocstr

numeric threshold for the minimum collostruction strength. By default it is set to 2.

Value

A tibble/data frame

References

Rajeg, G. P. W. (2019). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS. PhD Thesis. Monash University. Melbourne, Australia. https://doi.org/10.26180/5cac231a97fb1.

Examples

# for distinctive metaphors
mdca_res <- mdca(df = phd_data_metaphor,
                 cxn_var = "synonyms",
                 coll_var = "metaphors",
                 correct_holm = TRUE,
                 concise_output = TRUE,
                 already_count_table = FALSE,
                 assocstr_digits = 3)
# retrieve distinctive/attracted metaphors for "kebahagiaan".
mdca_attr(df = mdca_res,
          filter_by = "cxn",
          cxn_var = "synonyms",
          cxn_type = "kebahagiaan",
          min_assocstr = 2)
#> # A tibble: 4 × 8
#>   metaphors             synonyms      n    exp assocstr p_binomial p_holm  dec  
#>   <chr>                 <chr>     <int>  <dbl>    <dbl> <chr>      <chr>   <chr>
#> 1 happiness is a desir… kebahagi…   125  55.2     20.2  6.190e-21  3.838e… ***  
#> 2 intensity of happine… kebahagi…    23  10.7      3.87 1.357e-04  8.032e… ms   
#> 3 happiness is a posse… kebahagi…   181 141.       3.77 1.711e-04  1.011e… ns   
#> 4 happiness is a movin… kebahagi…    19   9.98     2.58 2.650e-03  1.000e… ns