Repelled items from MDCA

A function to retrieve list of repelled items for a construction from the results of mdca.

mdca_repel(df = NULL, filter_by = "cxn", cxn_var = "synonyms",
  cxn_type = NULL, coll_var = NULL, coll_type = NULL, min_assocstr = -2)

Arguments

df
output data frame from link{mdca}.

filter_by
character string indicating whether the output of link{mdca} is filtered according to the construction variable (i.e. "cxn") or the collexemes/collocates variable (i.e. "colloc"). If it is left NULL (the default), the data is filtered according to the value specified in the min_assocstr argument.

cxn_var
character string for the column name of the construction variable (i.e., "synonyms").

cxn_type
character strings of regular expressions for the HAPPINESS synonyms.

coll_var
character string for the column name of the metaphor variable (i.e., "metaphors").

coll_type
character strings of regular expressions for the conceptual metaphors.

min_assocstr
numeric threshold for the minimum collostruction strength. By default it is set to -2.

Value

A tibble/data frame

Examples

# for distinctive metaphors
mdca_res <- mdca(df = phd_data_metaphor,
                 cxn_var = "synonyms",
                 coll_var = "metaphors",
                 correct_holm = TRUE,
                 concise_output = TRUE,
                 already_count_table = FALSE,
                 assocstr_digits = 3)
# retrieve the repelled metaphors for "kebahagiaan".
mdca_repel(df = mdca_res,
          filter_by = "cxn",
          cxn_var = "synonyms",
          cxn_type = "kebahagiaan",
          min_assocstr = -2)

#> # A tibble: 7 x 8
#>   metaphors             synonyms      n   exp assocstr p_binomial p_holm  dec  
#>   <chr>                 <chr>     <int> <dbl>    <dbl> <chr>      <chr>   <chr>
#> 1 happiness is an (un)… kebahagi…    14 39.7     -6.44 3.629e-07  2.203e… ***  
#> 2 happiness is a conta… kebahagi…    34 67.4     -6.10 7.959e-07  4.815e… ***  
#> 3 happiness is a subme… kebahagi…     5 18.6     -4.18 6.566e-05  3.913e… *    
#> 4 happiness is (un)mix… kebahagi…     3 10.5     -2.43 3.739e-03  1.000e… ns   
#> 5 happiness is a restr… kebahagi…     0  4.90    -2.36 4.410e-03  1.000e… ns   
#> 6 happiness is a liqui… kebahagi…    17 29.4     -2.29 5.171e-03  1.000e… ns   
#> 7 happiness is a locat… kebahagi…    20 31.8     -2.00 9.974e-03  1.000e… ns