The main metaphor data frame

This data is used for the token frequency, type frequency and type-per-token ratio analyses in Rajeg (2019, Ch. 5 and Ch. 6), and for the Multiple Distinctive Collexeme Analysis in Rajeg (2019, Ch. 7). The data consists of the following variables:

corpus

Name of the corpus files

sentence_id

Character strings of the sentence id (or line id for WebCorp and Sketch Engine data) in which the happiness synonyms are found in the corpus

synonyms

Character strings of the synonyms

m_expr

Metaphorical expressions containing the synonyms, that is, the metaphorical patterns

m_expr_gloss

English gloss of the metaphorical patterns

lu

Source-frame lexical units (henceforth, LUs) in the metaphorical patterns

lu_gloss

English gloss of the source-frame LUs

lu_phrase_type

Syntactic category of the source-frame LUs (e.g., verb, adjective)

voice_verbal_lu

Grammatical voice of the verbal, source-frame LUs in the metaphorical patterns

source_frames

The source frame categories evoked by the source-frame LUs in the metaphorical patterns

metaphors

Conceptual metaphors evoked by the metaphorical patterns. This is based on grouping similar frame-role-mapping of the target-frame words in the source-frames

submappings

Submappings of the metaphors

body_part_inclusion

Character strings indicating whether the metaphorical patterns include/mention body-part terms ("y" for 'yes'), person ("person"), others type ("others"), or none ("NA")

body_part_terms

Character strings of the body-part terms

sentence_match

The sentence-match from which the metaphorical patterns of the synonyms are extracted

phd_data_metaphor

Format

An object of class tbl_df (inherits from tbl, data.frame) with 3638 rows and 15 columns.

References

Rajeg, G. P. W. (2019). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS (PhD Thesis). Monash University. Melbourne, Australia. https://doi.org/10.26180/5cac231a97fb1