Retrieve the (top-n most frequent) Lexical Units of a metaphor

The function is designed to easily retrieve a tibble data frame of the top-n most frequent Lexical Units (LU) of the prominent metaphors discussed in Chapter 5 and 6.

get_lu_table(metaphor = "regular expressions", top_n_only = TRUE,
  top_n_limit = 10, df = NULL, submapping_perc = FALSE,
  incl_submappings = FALSE)

Arguments

metaphor
regular expressions for the relevant metaphor.

top_n_only
logical; the default is TRUE with the number specified in top_n_limit argument. If FALSE, it prints all the LU for the metaphor.

top_n_limit
integer; how many most frequent LU should be printed? The default is 10.

df
the data frame for the thesis (phd_data_metaphor). This argument is not put first because the main attention when using the function is on varying the metaphor (hence, coming first) rather than on varying the database.

submapping_perc
logical; wheter to include the percentages of the LU by submappings. If TRUE, the function only print out the percentages of the LU by submappings. To include the submapping types, specify the incl_submappings argument to TRUE as well (the default is FALSE).

incl_submappings
logical; whether to include the submappings inferred from the LU (TRUE) or not (FALSE -- the default).

Value

A tbl_df.

Examples

get_lu_table(metaphor = "destination", df = phd_data_metaphor)

#> # A tibble: 0 x 4
#> # ... with 4 variables: Lexical_units <chr>, Gloss <chr>, N <int>,
#> #   Perc_overall <dbl>