Select a (set of) frequency profile column(s) for a metaphor

Select a (set of) frequency profile column(s) for a metaphor

get_meta_freq_profiles(metaphor = NULL, ...,
  metaphor_var = "metaphors", df = NULL)

Arguments

metaphor
character string of regular expressions for the metaphors.

...
any bare variable/column name of the frequency profiles from ttr function.

metaphor_var
character string for the column name of the metaphor variable (i.e., "metaphors").

df
data frame storing the results of ttr.

Value

A tibble data frame

References

Rajeg, G. P. W. (2018). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS (PhD Thesis). Monash University. Melbourne, Australia.

Examples

# perform ttr
ttr_metaphor <- ttr(phd_data_metaphor,
                    schema_var = "metaphors",
                    lexunit_var = "lu",
                    float_digits = 2)

# retrieve the profile
get_meta_freq_profiles("food", # regex for the metaphor
                       token, type_lu, type_per_token_lu, # inputs for the dot-dot-dot (...)
                       metaphor_var = "metaphors",
                       df = ttr_metaphor)

#> # A tibble: 1 x 4
#>   metaphors         token type_lu type_per_token_lu
#>   <chr>             <int>   <int>             <dbl>
#> 1 happiness is food   108      17              15.7