Retrieve the submappings of the metaphors overall or by synonyms/words

A utility function to get the descriptive data for the submappings of a metaphor (for the aggregated synonyms data or for a given synonym, if specified). This function is used in Rajeg's (2018, Chapter 5 & 6).

get_submappings(
  metaphor = NULL,
  word = NULL,
  df = NULL,
  metaphor_var = "metaphors",
  submet_var = "submappings",
  synonym_var = "synonyms",
  lexunit_var = "lu"
)

Arguments

metaphor

character strings of regular expressions for the metaphor.

word

character strings of regular expressions for the happiness synonyms.

df

the data frame (phd_data_metaphor).

metaphor_var

character string for the column name of the metaphor variable (i.e., "metaphors").

submet_var

character string for the column name of the submapping variable (i.e., "submappings").

synonym_var

character string for the column name of the synonyms variable (i.e., "synonyms").

lexunit_var

character string for the column name of the lexical unit variable (i.e., "lu").

Value

A tibble/data frame of the descriptive statistics of the submappings, namely the raw and relative (i.e. percentages of the) token and type frequencies.

References

Rajeg, G. P. W. (2018). Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS (PhD Thesis). Monash University. Melbourne, Australia.

Examples

submet <- get_submappings(metaphor = "possessable object$", df = phd_data_metaphor)